Renamed lib/api file to wpgraphql so serverless paths for next would route correctly
