+++ lib/wpGraphQL.js
